test: improve tests for unhandled rejections
